Added scenes for UI, Player and Gamestate, and a tool to quickly load into game + Extras
Created separate scenes for the persistent elements in the game, such as the UI, the player character and various managers. I've also added a menu shortcut to load into a scene with automatically added UI, Player Character and Game Managers. Ideally, this tool could be extended to allow developers to load into the game with any predefined gamestate they wish to test. Also, changed the enemy so that it grabs references in Start() instead of Awake(), as persistant scenes has not yet been loaded when Awake() is called.